requirements.txt
flask
werkzeug
pdfplumber
python-docx
scikit-learn
